scripts/automations/daily_news_curator/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.10
pyyaml>=6.0
